refresh cliplist on screen card
