+++ src/app/manage/screen/screen-overview/screen-overview.component.html
@@ -7,7 +7,7 @@
                        (onEditScreenClipOptions)="onClipOptions($event, screen)"
                        (onPreview)="onPreview($event, screen)"
                        (onReload)="onReload(screen)"
-                       [info]="screen"
+                       [screenId]="screen.id"
       >
 
       </app-screen-info>
