+++ src/app/manage/screen/screen-overview/screen-overview.component.ts
@@ -1,8 +1,8 @@
 import {Component, OnInit, TrackByFunction} from '@angular/core';
-import {Clip, HasId, Screen, ScreenViewEntry} from "@memebox/contracts";
+import {Clip, HasId, Screen} from "@memebox/contracts";
 import {Observable} from "rxjs";
 import {map} from "rxjs/operators";
-import {AppService, EXPRESS_BASE} from "../../../state/app.service";
+import {AppService} from "../../../state/app.service";
 import {AppQueries} from "../../../state/app.queries";
 import {DialogService} from "../../../shared/components/dialogs/dialog.service";
 import {WebsocketService} from "../../../core/services/websocket.service";
@@ -13,16 +13,6 @@ import {
   ClipAssigningMode
 } from "../../../shared/components/dialogs/clip-assigning-dialog/clip-assigning-dialog/clip-assigning-dialog.component";
 
-function createLocalOrProductionUrlBase() {
-  const port = location.port;
-  let urlBase = EXPRESS_BASE;
-
-  if (port === '4200') {
-    urlBase = location.host;
-  }
-
-  return urlBase;
-}
 
 function timeout(ms) {
   return new Promise(resolve => setTimeout(resolve, ms));
@@ -35,10 +25,9 @@ function timeout(ms) {
 })
 export class ScreenOverviewComponent implements OnInit {
 
-  public screenList: Observable<ScreenViewEntry[]> = this._queries.screensList$.pipe(
+  public screenList: Observable<Screen[]> = this._queries.screensList$.pipe(
     map(stateUrlArray => stateUrlArray.map(screen => ({
-      ...screen,
-      url: `${createLocalOrProductionUrlBase()}/#/screen/${screen.id}`
+      ...screen
     })))
   )
 
@@ -66,7 +55,7 @@ export class ScreenOverviewComponent implements OnInit {
     this.showDialog({});
   }
 
-  async delete(obsInfo: ScreenViewEntry) {
+  async delete(obsInfo: Screen) {
     const confirmationResult = await this._dialog.showConfirmationDialog(
       {
         title: 'Are you sure you want to delete this screen?'
@@ -94,7 +83,7 @@ export class ScreenOverviewComponent implements OnInit {
     )
   }
 
-  deleteAssigned(obsInfo: ScreenViewEntry, clipId: string) {
+  deleteAssigned(obsInfo: Screen, clipId: string) {
     this.service.deleteScreenClip(obsInfo.id, clipId);
   }
 
@@ -106,7 +95,7 @@ export class ScreenOverviewComponent implements OnInit {
     });
   }
 
-  async onPreview(clipId: string, screen: ScreenViewEntry) {
+  async onPreview(clipId: string, screen: Screen) {
     if (clipId) {
       this.webSocket.triggerClipOnScreen(clipId, screen.id);
     } else {
@@ -118,7 +107,7 @@ export class ScreenOverviewComponent implements OnInit {
 
   }
 
-  onReload(screen: ScreenViewEntry) {
+  onReload(screen: Screen) {
     this.webSocket.triggerReloadScreen(screen.id);
     this.snackbar.normal(`Screen: ${screen.name} reloaded`);
   }
